reorg to use base object & do renames
