+++ specparam/objs/__init__.py
@@ -1,6 +1,6 @@
 """Objects sub-module, for model objects and functions that operate on model objects."""
 
-from .fit import SpectralModel
+from .model import SpectralModel
 from .group import SpectralGroupModel
 from .utils import (compare_model_objs, average_group, average_reconstructions,
                     combine_model_objs, fit_models_3d)
