--- specparam/objs/bfit.py
@@ -1,212 +1,0 @@
-"""Define base fit model object."""
-
-import numpy as np
-
-from specparam.data import FitResults, ModelSettings
-from specparam.core.items import OBJ_DESC
-
-###################################################################################################
-###################################################################################################
-
-class BaseFit():
-    """   """
-
-    def __init__(self, aperiodic_mode, periodic_mode, debug_mode=False, verbose=True):
-
-        self.aperiodic_mode = aperiodic_mode
-        self.periodic_mode = periodic_mode
-        self.set_debug_mode(debug_mode)
-        self.verbose = verbose
-
-
-    def fit(self):
-        raise NotImplementedError('The method needs to overloaded with a fit procedure!')
-
-
-    def get_settings(self):
-        """Return user defined settings of the current object.
-
-        Returns
-        -------
-        ModelSettings
-            Object containing the settings from the current object.
-        """
-
-        return ModelSettings(**{key : getattr(self, key) \
-                             for key in OBJ_DESC['settings']})
-
-
-    def get_results(self):
-        """Return model fit parameters and goodness of fit metrics.
-
-        Returns
-        -------
-        FitResults
-            Object containing the model fit results from the current object.
-        """
-
-        return FitResults(**{key.strip('_') : getattr(self, key) \
-            for key in OBJ_DESC['results']})
-
-
-    def set_debug_mode(self, debug):
-        """Set debug mode, which controls if an error is raised if model fitting is unsuccessful.
-
-        Parameters
-        ----------
-        debug : bool
-            Whether to run in debug mode.
-        """
-
-        self._debug = debug
-
-
-    def _reset_results(self, clear_results=False):
-        """Set, or reset, results attributes to empty.
-
-        Parameters
-        ----------
-        clear_results : bool, optional, default: False
-            Whether to clear model results attributes.
-        """
-
-        if clear_results:
-
-            # Aperiodic parameers
-            self.aperiodic_params_ = np.nan
-
-            # Periodic parameters
-            self.gaussian_params_ = np.nan
-            self.peak_params_ = np.nan
-
-            # Note - for ap / pe params, move to something like `xx_params` and `_xx_params`
-
-            # Goodness of fit measures
-            self.r_squared_ = np.nan
-            self.error_ = np.nan
-            # Note: move to `self.gof` or similar
-
-            # Modeled spectrum components
-            self.modeled_spectrum_ = None
-            self._spectrum_flat = None
-            self._spectrum_peak_rm = None
-            self._ap_fit = None
-            self._peak_fit = None
-
-
-    def _calc_r_squared(self):
-        """Calculate the r-squared goodness of fit of the model, compared to the original data."""
-
-        r_val = np.corrcoef(self.power_spectrum, self.modeled_spectrum_)
-        self.r_squared_ = r_val[0][1] ** 2
-
-
-    def _calc_error(self, metric=None):
-        """Calculate the overall error of the model fit, compared to the original data.
-
-        Parameters
-        ----------
-        metric : {'MAE', 'MSE', 'RMSE'}, optional
-            Which error measure to calculate:
-            * 'MAE' : mean absolute error
-            * 'MSE' : mean squared error
-            * 'RMSE' : root mean squared error
-
-        Raises
-        ------
-        ValueError
-            If the requested error metric is not understood.
-
-        Notes
-        -----
-        Which measure is applied is by default controlled by the `_error_metric` attribute.
-        """
-
-        # If metric is not specified, use the default approach
-        metric = self._error_metric if not metric else metric
-
-        if metric == 'MAE':
-            self.error_ = np.abs(self.power_spectrum - self.modeled_spectrum_).mean()
-
-        elif metric == 'MSE':
-            self.error_ = ((self.power_spectrum - self.modeled_spectrum_) ** 2).mean()
-
-        elif metric == 'RMSE':
-            self.error_ = np.sqrt(((self.power_spectrum - self.modeled_spectrum_) ** 2).mean())
-
-        else:
-            error_msg = "Error metric '{}' not understood or not implemented.".format(metric)
-            raise ValueError(error_msg)
-
-
-    def _add_from_dict(self, data):
-        """Add data to object from a dictionary.
-
-        Parameters
-        ----------
-        data : dict
-            Dictionary of data to add to self.
-        """
-
-        # Reconstruct object from loaded data
-        for key in data.keys():
-            setattr(self, key, data[key])
-
-
-class BaseFit2D(BaseFit):
-
-    def __init__(self, aperiodic_mode, periodic_mode, debug_mode=False, verbose=True):
-
-        BaseFit.__init__(self, aperiodic_mode, periodic_mode, debug_mode=False, verbose=True)
-
-        self._reset_group_results()
-
-
-    def __len__(self):
-        """Define the length of the object as the number of model fit results available."""
-
-        return len(self.group_results)
-
-
-    def __iter__(self):
-        """Allow for iterating across the object by stepping across model fit results."""
-
-        for result in self.group_results:
-            yield result
-
-
-    def __getitem__(self, index):
-        """Allow for indexing into the object to select model fit results."""
-
-        return self.group_results[index]
-
-
-    def _reset_group_results(self, length=0):
-        """Set, or reset, results to be empty.
-
-        Parameters
-        ----------
-        length : int, optional, default: 0
-            Length of list of empty lists to initialize. If 0, creates a single empty list.
-        """
-
-        self.group_results = [[]] * length
-
-
-    @property
-    def has_model(self):
-        """Indicator for if the object contains model fits."""
-
-        return True if self.group_results else False
-
-
-    def get_results(self):
-        """Return the results run across a group of power spectra."""
-
-        return self.group_results
-
-
-    def _get_results(self):
-        """Create an alias to SpectralModel.get_results for the group object, for internal use."""
-
-        return super().get_results()
