+++ specparam/tests/objs/test_base.py
@@ -2,27 +2,19 @@
 
 from specparam.sim import sim_power_spectrum
 
-# TEMP:
-from specparam.objs.bfit import BaseFit
-from specparam.objs.data import BaseData
+from specparam.objs.algorithm import SpectralFitAlgorithm
 
 from specparam.objs.base import *
 
 ###################################################################################################
 ###################################################################################################
 
-# def test_base_object():
-#     """Check base object initializes properly."""
+def test_base_object():
 
-#     assert BaseSpectralModel()
-
-def test_base_fit():
-
-    class TestBase(BaseSpectralModel, BaseFit, BaseData):
+    class TestBase(SpectralFitAlgorithm, BaseObject):
         def __init__(self):
-            BaseData.__init__(self)
-            BaseFit.__init__(self, aperiodic_mode='fixed', periodic_mode='gaussian')
-            BaseSpectralModel.__init__(self)
+            BaseObject.__init__(self, aperiodic_mode='fixed', periodic_mode='gaussian')
+            SpectralFitAlgorithm.__init__(self)
 
     tbase = TestBase()
     tbase.fit(*sim_power_spectrum([3, 50], [50, 2], [10, 0.5, 2, 20, 0.3, 4]))
